commit_8

+++ work_with_database/phones/models.py
@@ -5,7 +5,7 @@ from django.utils.text import slugify
 class Phone(models.Model):
     # TODO: Добавьте требуемые поля
     name = models.CharField(max_length=150)
-    prise = models.FloatField()
+    prise = models.IntegerField()
     release_date = models.DateField()
     lte_exists = models.BooleanField()
     slag = models.SlugField(max_length=150, name = slugify(name))
